third_party/WebKit/LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-ax-aria-attributes-expected.txt

This test checks aria-help attribute of fields in multiple fields date input UI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial empty state
PASS focusedElementDescription() is "Month, AXValueDescription: mm, intValue:0, range:1-12"
PASS focusedElementDescription() is "Day, AXValueDescription: dd, intValue:0, range:1-31"
PASS focusedElementDescription() is "Year, AXValueDescription: yyyy, intValue:0, range:1-275760"
Non-empty value
PASS focusedElementDescription() is "Month, AXValueDescription: 10, intValue:10, range:1-12"
PASS focusedElementDescription() is "Day, AXValueDescription: 09, intValue:9, range:1-31"
PASS focusedElementDescription() is "Year, AXValueDescription: 2012, intValue:2012, range:1-275760"
Empty value
PASS focusedElementDescription() is "Year, AXValueDescription: yyyy, intValue:0, range:1-275760"
PASS focusedElementDescription() is "Day, AXValueDescription: dd, intValue:0, range:1-31"
PASS focusedElementDescription() is "Month, AXValueDescription: mm, intValue:0, range:1-12"

PASS successfullyParsed is true

TEST COMPLETE
